Move the counters to a dedicated file
